Added testers for Moisture,GasSense, and RelayX1. Fixed minor analog read error in Moisture and GasSense.

+++ CPP/Moisture/Moisture.h
@@ -16,7 +16,7 @@ namespace GHI
 			public:
 				Moisture(int socket);
 
-				int GetMoistureReading(int samples);
+				double GetMoistureReading(int samples);
 		};
 	}
 }
